+++ packages/crayons-i18n/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.3.0-beta.4](https://github.com/freshworks/crayons/compare/@freshworks/crayons-i18n@4.3.0-beta.3...@freshworks/crayons-i18n@4.3.0-beta.4) (2025-01-29)
+
+### Bug Fixes
+
+- naming convention changes for SLA timer ([85ba1b7](https://github.com/freshworks/crayons/commit/85ba1b7f8884967b50c7c5443f050d2c6f7c78ce))
+
+## [4.3.0-beta.3](https://github.com/freshworks/crayons/compare/@freshworks/crayons-i18n@4.3.0-beta.2...@freshworks/crayons-i18n@4.3.0-beta.3) (2024-12-10)
+
+### Bug Fixes
+
+- added the error message for section inside section ([#959](https://github.com/freshworks/crayons/issues/959)) ([277a49f](https://github.com/freshworks/crayons/commit/277a49ff3ca19a659f84d914712448e83e5179a8))
+
 ## [4.3.0-beta.2](https://github.com/freshworks/crayons/compare/@freshworks/crayons-i18n@4.3.0-beta.1...@freshworks/crayons-i18n@4.3.0-beta.2) (2024-10-15)
 
 ### Bug Fixes